refactor: moved sockets spec str parsing in CLI, separate perf and powercap modules
