+++ sources/rapl/src/domain_type.rs
